// client/src/DTOs/CardDto.ts
export interface CardDto {
    // cardId: number;
    // ownerId: number;
    ownerLogin: string;
    createdDate: string;
    expirationDate: string;
    isActive: boolean;
    abilityChoice: number;
    abilityKeyWords: string;
    locationChoice: number;
    locationCountry: string;
    locationRegion: string;
    languageChoice: string;
    ageChoice: number;
}